test: move tests to TransactionIntegrationTest and make them all pass
