+++ sdk/src/integrationTest/java/FileAppendIntegrationTest.java
@@ -188,115 +188,4 @@ public class FileAppendIntegrationTest {
 
         testEnv.close();
     }
-
-    @Test
-    void canSerializeDeserializeExecute() throws Exception {
-        // There are potential bugs in FileAppendTransaction which require more than one node to trigger.
-        var testEnv = new IntegrationTestEnv(2);
-
-        // Skip if using local node.
-        // Note: this check should be removed once the local node is supporting multiple nodes.
-        testEnv.assumeNotLocalNode();
-
-        var response = new FileCreateTransaction()
-            .setKeys(testEnv.operatorKey)
-            .setContents("[e2e::FileCreateTransaction]")
-            .execute(testEnv.client);
-
-        var fileId = Objects.requireNonNull(response.getReceipt(testEnv.client).fileId);
-
-        Thread.sleep(5000);
-
-        @Var var info = new FileInfoQuery()
-            .setFileId(fileId)
-            .execute(testEnv.client);
-
-        assertThat(info.fileId).isEqualTo(fileId);
-        assertThat(info.size).isEqualTo(28);
-        assertThat(info.isDeleted).isFalse();
-        assertThat(info.keys).isNotNull();
-        assertThat(info.keys.getThreshold()).isNull();
-        assertThat(info.keys).isEqualTo(KeyList.of(testEnv.operatorKey));
-
-        var fileAppendTransaction = new FileAppendTransaction()
-            .setFileId(fileId)
-            .setContents(Contents.BIG_CONTENTS);
-
-        var transactionBytesSerialized = fileAppendTransaction.toBytes();
-        FileAppendTransaction fileAppendTransactionDeserialized = (FileAppendTransaction) Transaction.fromBytes(transactionBytesSerialized);
-
-        fileAppendTransactionDeserialized
-            .execute(testEnv.client)
-            .getReceipt(testEnv.client);
-
-        var contents = new FileContentsQuery()
-            .setFileId(fileId)
-            .execute(testEnv.client);
-
-        assertThat(contents.toStringUtf8()).isEqualTo("[e2e::FileCreateTransaction]" + Contents.BIG_CONTENTS);
-
-        info = new FileInfoQuery()
-            .setFileId(fileId)
-            .execute(testEnv.client);
-
-        assertThat(info.fileId).isEqualTo(fileId);
-        assertThat(info.size).isEqualTo(13522);
-        assertThat(info.isDeleted).isFalse();
-        assertThat(info.keys).isNotNull();
-        assertThat(info.keys.getThreshold()).isNull();
-        assertThat(info.keys).isEqualTo(KeyList.of(testEnv.operatorKey));
-
-        new FileDeleteTransaction()
-            .setFileId(fileId)
-            .execute(testEnv.client)
-            .getReceipt(testEnv.client);
-
-        testEnv.close();
-    }
-
-    @Test
-    void canFreezeSignSerializeDeserializeAndCompare() throws Exception {
-        // There are potential bugs in FileAppendTransaction which require more than one node to trigger.
-        var testEnv = new IntegrationTestEnv(2);
-
-        // Skip if using local node.
-        // Note: this check should be removed once the local node is supporting multiple nodes.
-        testEnv.assumeNotLocalNode();
-
-        var privateKey = PrivateKey.generateED25519();
-
-        var response = new FileCreateTransaction()
-            .setKeys(testEnv.operatorKey)
-            .setContents("[e2e::FileCreateTransaction]")
-            .execute(testEnv.client);
-
-        var fileId = Objects.requireNonNull(response.getReceipt(testEnv.client).fileId);
-
-        Thread.sleep(5000);
-
-        @Var var info = new FileInfoQuery()
-            .setFileId(fileId)
-            .execute(testEnv.client);
-
-        assertThat(info.fileId).isEqualTo(fileId);
-        assertThat(info.size).isEqualTo(28);
-        assertThat(info.isDeleted).isFalse();
-        assertThat(info.keys).isNotNull();
-        assertThat(info.keys.getThreshold()).isNull();
-        assertThat(info.keys).isEqualTo(KeyList.of(testEnv.operatorKey));
-
-        var fileAppendTransaction = new FileAppendTransaction()
-            .setFileId(fileId)
-            .setContents(Contents.BIG_CONTENTS)
-            .freezeWith(testEnv.client)
-            .sign(privateKey);
-
-        var transactionBytesSerialized = fileAppendTransaction.toBytes();
-        FileAppendTransaction fileAppendTransactionDeserialized = (FileAppendTransaction) Transaction.fromBytes(transactionBytesSerialized);
-
-        var transactionBytesReserialized = fileAppendTransactionDeserialized.toBytes();
-        assertThat(transactionBytesSerialized).isEqualTo(transactionBytesReserialized);
-
-        testEnv.close();
-    }
 }
