+++ sdk/src/integrationTest/java/TransactionIntegrationTest.java
@@ -3,6 +3,13 @@ import com.google.protobuf.ByteString;
 import com.hedera.hashgraph.sdk.AccountCreateTransaction;
 import com.hedera.hashgraph.sdk.AccountDeleteTransaction;
 import com.hedera.hashgraph.sdk.AccountId;
+import com.hedera.hashgraph.sdk.FileAppendTransaction;
+import com.hedera.hashgraph.sdk.FileContentsQuery;
+import com.hedera.hashgraph.sdk.FileCreateTransaction;
+import com.hedera.hashgraph.sdk.FileDeleteTransaction;
+import com.hedera.hashgraph.sdk.FileInfoQuery;
+import com.hedera.hashgraph.sdk.Hbar;
+import com.hedera.hashgraph.sdk.KeyList;
 import com.hedera.hashgraph.sdk.PrivateKey;
 import com.hedera.hashgraph.sdk.Transaction;
 import com.hedera.hashgraph.sdk.TransactionId;
@@ -19,6 +26,7 @@ import com.hedera.hashgraph.sdk.proto.TransactionBody;
 import com.hedera.hashgraph.sdk.proto.TransactionID;
 import com.hedera.hashgraph.sdk.proto.TransactionList;
 import com.hedera.hashgraph.sdk.proto.TransferList;
+import java.util.Objects;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -26,9 +34,11 @@ import org.junit.jupiter.api.Test;
 import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 public class TransactionIntegrationTest {
+
     @Test
     @DisplayName("transaction hash in transaction record is equal to the derived transaction hash")
     void transactionHashInTransactionRecordIsEqualToTheDerivedTransactionHash() throws Exception {
@@ -62,8 +72,176 @@ public class TransactionIntegrationTest {
     }
 
     @Test
-    @DisplayName("transaction can be serialized into bytes, deserialized, signature added and executed")
-    void transactionFromToBytes() throws Exception {
+    @DisplayName("transaction can be serialized into bytes, deserialized and be equal to the original one")
+    void canSerializeDeserializeCompareFields() throws Exception {
+        var testEnv = new IntegrationTestEnv(1);
+
+        var adminKey = PrivateKey.generateECDSA();
+        var publicKey = adminKey.getPublicKey();
+
+        var createAccountTransaction = new AccountCreateTransaction()
+            .setKey(publicKey)
+            .setInitialBalance(new Hbar(5L));
+
+        var expectedNodeAccountIds = createAccountTransaction.getNodeAccountIds();
+        var expectedBalance = new Hbar(5L);
+
+        var transactionBytesSerialized = createAccountTransaction.toBytes();
+        AccountCreateTransaction createAccountTransactionDeserialized = (AccountCreateTransaction) Transaction.fromBytes(transactionBytesSerialized);
+
+        assertThat(expectedNodeAccountIds).isEqualTo(createAccountTransactionDeserialized.getNodeAccountIds());
+        assertThat(expectedBalance).isEqualTo(createAccountTransactionDeserialized.getInitialBalance());
+        assertThatExceptionOfType(IllegalStateException.class).isThrownBy(
+            createAccountTransactionDeserialized::getTransactionId);
+
+        testEnv.close();
+    }
+
+    @Test
+    @DisplayName("incomplete transaction can be serialized into bytes, deserialized, signature added and executed")
+    void canIncompleteSerializeDeserializeAndExecute() throws Exception {
+        var testEnv = new IntegrationTestEnv(1);
+
+        var adminKey = PrivateKey.generateECDSA();
+        var publicKey = adminKey.getPublicKey();
+
+        var createAccountTransaction = new AccountCreateTransaction()
+            .setKey(publicKey)
+            .setInitialBalance(new Hbar(5L));
+
+        var transactionBytesSerialized = createAccountTransaction.toBytes();
+        AccountCreateTransaction createAccountTransactionDeserialized = (AccountCreateTransaction) Transaction.fromBytes(transactionBytesSerialized);
+
+        var txReceipt = createAccountTransactionDeserialized
+            .execute(testEnv.client)
+            .getReceipt(testEnv.client);
+
+        new AccountDeleteTransaction()
+            .setAccountId(txReceipt.accountId)
+            .setTransferAccountId(testEnv.client.getOperatorAccountId())
+            .freezeWith(testEnv.client)
+            .sign(adminKey)
+            .execute(testEnv.client);
+    }
+
+    @Test
+    @DisplayName("transaction can be serialized into bytes, deserialized, edited and executed")
+    void canSerializeDeserializeEditExecuteCompareFields() throws Exception {
+        var testEnv = new IntegrationTestEnv(1);
+
+        var adminKey = PrivateKey.generateECDSA();
+        var publicKey = adminKey.getPublicKey();
+
+        var createAccountTransaction = new AccountCreateTransaction()
+            .setKey(publicKey);
+
+        var expectedBalance = new Hbar(5L);
+        var nodeAccountIds = testEnv.client.getNetwork().values().stream().toList();
+
+        var transactionBytesSerialized = createAccountTransaction.toBytes();
+        AccountCreateTransaction createAccountTransactionDeserialized = (AccountCreateTransaction) Transaction.fromBytes(transactionBytesSerialized);
+
+        var txReceipt = createAccountTransactionDeserialized
+            .setInitialBalance(new Hbar(5L))
+            .setNodeAccountIds(nodeAccountIds)
+            .setTransactionId(TransactionId.generate(testEnv.client.getOperatorAccountId()))
+            .execute(testEnv.client)
+            .getReceipt(testEnv.client);
+
+        assertThat(expectedBalance).isEqualTo(createAccountTransactionDeserialized.getInitialBalance());
+
+        new AccountDeleteTransaction()
+            .setAccountId(txReceipt.accountId)
+            .setTransferAccountId(testEnv.client.getOperatorAccountId())
+            .freezeWith(testEnv.client)
+            .sign(adminKey)
+            .execute(testEnv.client);
+    }
+
+    @Test
+    @DisplayName("transaction can be serialized into bytes, deserialized, edited and executed")
+    void canSerializeDeserializeEditExecuteCompareFields_2() throws Exception {
+        var testEnv = new IntegrationTestEnv(1);
+
+        var adminKey = PrivateKey.generateECDSA();
+        var publicKey = adminKey.getPublicKey();
+
+        var nodeAccountIds = testEnv.client.getNetwork().values().stream().toList();
+
+        var createAccountTransaction = new AccountCreateTransaction()
+            .setKey(publicKey)
+            .setNodeAccountIds(nodeAccountIds);
+
+        var expectedBalance = new Hbar(5L);
+
+        var transactionBytesSerialized = createAccountTransaction.toBytes();
+        AccountCreateTransaction createAccountTransactionDeserialized = (AccountCreateTransaction) Transaction.fromBytes(transactionBytesSerialized);
+
+        var txReceipt = createAccountTransactionDeserialized
+            .setInitialBalance(new Hbar(5L))
+            .setTransactionId(TransactionId.generate(testEnv.client.getOperatorAccountId()))
+            .execute(testEnv.client)
+            .getReceipt(testEnv.client);
+
+        assertThat(expectedBalance).isEqualTo(createAccountTransactionDeserialized.getInitialBalance());
+
+        new AccountDeleteTransaction()
+            .setAccountId(txReceipt.accountId)
+            .setTransferAccountId(testEnv.client.getOperatorAccountId())
+            .freezeWith(testEnv.client)
+            .sign(adminKey)
+            .execute(testEnv.client);
+    }
+
+    @Test
+    @DisplayName("complete frozen and signed transaction can be serialized into bytes, deserialized and executed")
+    void canFreezeSignSerializeDeserializeAndExecute() throws Exception {
+        var testEnv = new IntegrationTestEnv(1);
+
+        var adminKey = PrivateKey.generateECDSA();
+        var publicKey = adminKey.getPublicKey();
+
+        var evmAddress = publicKey.toEvmAddress();
+        var initialBalance = new Hbar(5L);
+        var autoRenewPeriod = java.time.Duration.ofSeconds(2592000);
+        var memo = "test account memo";
+        var maxAutomaticTokenAssociations = 4;
+
+        var createAccountTransaction = new AccountCreateTransaction()
+            .setKey(publicKey)
+            .setInitialBalance(initialBalance)
+            .setReceiverSignatureRequired(true)
+            .setAutoRenewPeriod(autoRenewPeriod)
+            .setAccountMemo(memo)
+            .setMaxAutomaticTokenAssociations(maxAutomaticTokenAssociations)
+            .setDeclineStakingReward(true)
+            .setAlias(evmAddress)
+            .freezeWith(testEnv.client)
+            .sign(adminKey);
+
+        var transactionBytesSerialized = createAccountTransaction.toBytes();
+        AccountCreateTransaction createAccountTransactionDeserialized = (AccountCreateTransaction) Transaction.fromBytes(transactionBytesSerialized);
+
+        var transactionBytesReserialized = createAccountTransactionDeserialized.toBytes();
+        assertThat(transactionBytesSerialized).isEqualTo(transactionBytesReserialized);
+
+        AccountCreateTransaction createAccountTransactionDeserialized2 = (AccountCreateTransaction) Transaction.fromBytes(transactionBytesReserialized);
+
+        var txResponse = createAccountTransactionDeserialized2.execute(testEnv.client);
+
+        var accountId = txResponse.getReceipt(testEnv.client).accountId;
+
+        new AccountDeleteTransaction()
+            .setAccountId(accountId)
+            .setTransferAccountId(testEnv.client.getOperatorAccountId())
+            .freezeWith(testEnv.client)
+            .sign(adminKey)
+            .execute(testEnv.client);
+    }
+
+    @Test
+    @DisplayName("complete frozen transaction can be serialized into bytes, deserialized, signature added and executed")
+    void canFreezeSerializeDeserializeAddSignatureAndExecute() throws Exception {
         var testEnv = new IntegrationTestEnv(1);
 
         var key = PrivateKey.generateED25519();
@@ -89,25 +267,144 @@ public class TransactionIntegrationTest {
             .setTransferAccountId(testEnv.operatorId)
             .freezeWith(testEnv.client);
 
+        System.out.println("TransactionId (before)= " + deleteTransaction.getTransactionId());
+        System.out.println("NodeAccountIds (before)= " + deleteTransaction.getNodeAccountIds());
         var updateBytes = deleteTransaction.toBytes();
 
         var sig1 = key.signTransaction(deleteTransaction);
 
         var deleteTransaction2 = Transaction.fromBytes(updateBytes);
 
+        System.out.println("TransactionId (after)= " + deleteTransaction2.getTransactionId());
+        System.out.println("NodeAccountIds (after)= " + deleteTransaction2.getNodeAccountIds());
+
         response = deleteTransaction2
             .addSignature(key.getPublicKey(), sig1)
             .execute(testEnv.client);
 
-        response.getReceipt(testEnv.client);
+        System.out.println("record= " + response.getRecord(testEnv.client));
+        System.out.println("receipt= " + response.getReceipt(testEnv.client));
+
+        testEnv.close();
+    }
+
+    @Test
+    @DisplayName("transaction can be serialized into bytes, deserialized, edited and executed")
+    void canSerializeDeserializeExecute() throws Exception {
+        // There are potential bugs in FileAppendTransaction which require more than one node to trigger.
+        var testEnv = new IntegrationTestEnv(2);
+
+        // Skip if using local node.
+        // Note: this check should be removed once the local node is supporting multiple nodes.
+        testEnv.assumeNotLocalNode();
+
+        var response = new FileCreateTransaction()
+            .setKeys(testEnv.operatorKey)
+            .setContents("[e2e::FileCreateTransaction]")
+            .execute(testEnv.client);
+
+        var fileId = Objects.requireNonNull(response.getReceipt(testEnv.client).fileId);
+
+        Thread.sleep(5000);
+
+        @Var var info = new FileInfoQuery()
+            .setFileId(fileId)
+            .execute(testEnv.client);
+
+        assertThat(info.fileId).isEqualTo(fileId);
+        assertThat(info.size).isEqualTo(28);
+        assertThat(info.isDeleted).isFalse();
+        assertThat(info.keys).isNotNull();
+        assertThat(info.keys.getThreshold()).isNull();
+        assertThat(info.keys).isEqualTo(KeyList.of(testEnv.operatorKey));
+
+        var fileAppendTransaction = new FileAppendTransaction()
+            .setFileId(fileId)
+            .setContents(Contents.BIG_CONTENTS);
+
+        var transactionBytesSerialized = fileAppendTransaction.toBytes();
+        FileAppendTransaction fileAppendTransactionDeserialized = (FileAppendTransaction) Transaction.fromBytes(transactionBytesSerialized);
+
+        fileAppendTransactionDeserialized
+            .execute(testEnv.client)
+            .getReceipt(testEnv.client);
+
+        var contents = new FileContentsQuery()
+            .setFileId(fileId)
+            .execute(testEnv.client);
+
+        assertThat(contents.toStringUtf8()).isEqualTo("[e2e::FileCreateTransaction]" + Contents.BIG_CONTENTS);
+
+        info = new FileInfoQuery()
+            .setFileId(fileId)
+            .execute(testEnv.client);
+
+        assertThat(info.fileId).isEqualTo(fileId);
+        assertThat(info.size).isEqualTo(13522);
+        assertThat(info.isDeleted).isFalse();
+        assertThat(info.keys).isNotNull();
+        assertThat(info.keys.getThreshold()).isNull();
+        assertThat(info.keys).isEqualTo(KeyList.of(testEnv.operatorKey));
+
+        new FileDeleteTransaction()
+            .setFileId(fileId)
+            .execute(testEnv.client)
+            .getReceipt(testEnv.client);
+
+        testEnv.close();
+    }
+
+    @Test
+    @DisplayName("transaction can be frozen, signed, serialized into bytes, deserialized and be equal to the original one")
+    void canFreezeSignSerializeDeserializeAndCompare() throws Exception {
+        // There are potential bugs in FileAppendTransaction which require more than one node to trigger.
+        var testEnv = new IntegrationTestEnv(2);
+
+        // Skip if using local node.
+        // Note: this check should be removed once the local node is supporting multiple nodes.
+        testEnv.assumeNotLocalNode();
+
+        var privateKey = PrivateKey.generateED25519();
+
+        var response = new FileCreateTransaction()
+            .setKeys(testEnv.operatorKey)
+            .setContents("[e2e::FileCreateTransaction]")
+            .execute(testEnv.client);
+
+        var fileId = Objects.requireNonNull(response.getReceipt(testEnv.client).fileId);
+
+        Thread.sleep(5000);
+
+        @Var var info = new FileInfoQuery()
+            .setFileId(fileId)
+            .execute(testEnv.client);
+
+        assertThat(info.fileId).isEqualTo(fileId);
+        assertThat(info.size).isEqualTo(28);
+        assertThat(info.isDeleted).isFalse();
+        assertThat(info.keys).isNotNull();
+        assertThat(info.keys.getThreshold()).isNull();
+        assertThat(info.keys).isEqualTo(KeyList.of(testEnv.operatorKey));
+
+        var fileAppendTransaction = new FileAppendTransaction()
+            .setFileId(fileId)
+            .setContents(Contents.BIG_CONTENTS)
+            .freezeWith(testEnv.client)
+            .sign(privateKey);
+
+        var transactionBytesSerialized = fileAppendTransaction.toBytes();
+        FileAppendTransaction fileAppendTransactionDeserialized = (FileAppendTransaction) Transaction.fromBytes(transactionBytesSerialized);
+
+        var transactionBytesReserialized = fileAppendTransactionDeserialized.toBytes();
+        assertThat(transactionBytesSerialized).isEqualTo(transactionBytesReserialized);
 
         testEnv.close();
     }
 
 
+
     // TODO: this test has a bunch of things hard-coded into it, which is kinda dumb, but it's a good idea for a test.
     //       Any way to fix it and bring it back?
-
     @Disabled
     @Test
     @DisplayName("transaction can be serialized into bytes, deserialized, signature added and executed")
